# backend/swagger.yaml
openapi: 3.0.3
info:
  title: NGO Helper API
  description: |
    A comprehensive REST API for NGO management, reporting, and analytics. 
    This API provides endpoints for authentication, report submission, bulk CSV uploads, 
    contact management, and admin dashboard functionality.
  version: 1.0.0
  contact:
    name: NGO Helper Team
    email: support@ngohelper.com
servers:
  - url: ${process.env.REACT_APP_API_URL}
    description: Development server
  - url: https://api.ngohelper.com
    description: Production server
...
paths:
...
  /api/report:
    post:
      tags:
        - Reports
      summary: Submit single report
      description: Submit a monthly NGO report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportSubmission'
      responses:
        '201':
          description: Report submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Report submitted successfully"
                  data:
                    $ref: '#/components/schemas/Report'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/contacts:
    post:
      tags:
        - Contacts
      summary: Submit contact form
      description: Submit a contact form inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactSubmission'
      responses:
        '201':
          description: Contact form submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Contact form submitted successfully"
                  data:
                    $ref: '#/components/schemas/Contact'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    get:
      tags:
        - Contacts
      summary: Get all contacts
      description: Get all contact form submissions (Admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Contacts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contact'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    ReportSubmission:
      type: object
      required:
        - ngoId
        - month
        - peopleHelped
        - eventsConducted
        - fundsUtilized
      properties:
        ngoId:
          type: string
          minLength: 1
          maxLength: 50
          description: NGO identifier
          example: "NGO001"
        month:
          type: string
          enum: [January, February, March, April, May, June, July, August, September, October, November, December]
          description: Month of the report
          example: "January"
        year:
          type: integer
          minimum: 2020
          maximum: 2030
          description: Year of the report (defaults to current year)
          example: 2024
        peopleHelped:
          type: integer
          minimum: 0
          description: Number of people helped
          example: 150
        eventsConducted:
          type: integer
          minimum: 0
          description: Number of events conducted
          example: 5
        fundsUtilized:
          type: number
          minimum: 0
          description: Amount of funds utilized in rupees
          example: 25000
    Report:
      type: object
      properties:
        id:
          type: string
          example: "report-123"
        ngo_id:
          type: string
          example: "NGO001"
        ngo_name:
          type: string
          example: "Hope Foundation"
        month:
          type: string
          example: "January"
        year:
          type: integer
          example: 2024
        people_helped:
          type: integer
          example: 150
        events_conducted:
          type: integer
          example: 5
        funds_utilized:
          type: number
          example: 25000
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
...
    ContactSubmission:
      type: object
      required:
        - name
        - email
        - subject
        - message
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Full name of the contact
          example: "John Doe"
        email:
          type: string
          format: email
          description: Email address of the contact
          example: "john.doe@example.com"
        ngoId:
          type: string
          nullable: true
          description: NGO ID if applicable
          example: "NGO001"
        subject:
          type: string
          enum: [Technical Support, Report Issues, Account Help, Feature Request, General Inquiry, Other]
          description: Subject of the inquiry
          example: "Technical Support"
        message:
          type: string
          minLength: 10
          maxLength: 1000
          description: Detailed message
          example: "I'm having trouble uploading my CSV file. Can you help?"
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        ngo_id:
          type: string
          nullable: true
          example: "NGO001"
        subject:
          type: string
          example: "Technical Support"
        message:
          type: string
          example: "I'm having trouble uploading my CSV file. Can you help?"
        status:
          type: string
          enum: [new, in_progress, resolved, closed]
          example: "new"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error message describing what went wrong"
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation failed"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Email is required"